=act split FSM java api in a separate file.
